requirements.txt
google-generativeai>=0.8.0
pdf2image>=1.17.0
Pillow>=10.0.0
python-dotenv>=1.0.0